pyproject.toml
[tool.poetry]
name = "automatic-umbrella"
version = "0.2.0"
description = ""
authors = ["John Paul Swaynos <jpswaynos@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.22.0"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.0.1"
